Console typeclass moved to typeclasses project, instances to instances project
